Moves Framework-related sweeper functions to separate package
